Crate ocaml_rs_smartptr

Re-exports

pub use inventory;

Modules

callable

func

ml_box

This module provides a safe wrapper around ocaml::Value to allow sending OCaml values between threads in Rust. The MlBox struct ensures that the OCaml runtime handle is available when recovering the value, maintaining safety.

ocaml_gen_extras

This module provides additional utilities and extensions for generating OCaml bindings.

ptr

This module provides the DynBox smart pointer, which is a wrapper around the registry’s DynArc with PhantomData for type safety. DynBox allows the user to wrap the object in a Mutex or shared RwLock. By default, using .into() will create a Mutex-protected version (exclusive).

registry

This module provides a Registry for managing type coercions and type information in a Rust-OCaml binding context. The Registry allows for the registration of coercion functions that convert between different types, leveraging Rust’s type system and OCaml’s garbage collector.

stubs

This module provides some stubs for OCaml (extern "C" functions). Not to be used from Rust.

Macros

ocaml_export

ocaml_gen_bindings

register_rtti

register_trait

register_type