ocaml_rs_smartptr::callable

Trait Callable

pub trait Callable<Ret>
where
    Ret: FromValue + OCamlDesc,
{
    // Required methods
    fn call_with(&self, gc: &Runtime, func: Value) -> Ret;
    fn describe_args(env: &Env, generics: &[&str]) -> Vec<String>;
    fn unique_id_args() -> Vec<u128>;

    // Provided methods
    fn ocaml_desc(env: &Env, generics: &[&str]) -> String { ... }
    fn unique_id() -> u128 { ... }
    fn process_result(&self, res: Result<Value, Error>) -> Ret { ... }
}

The Callable trait represents a function or closure that can be called with a set of arguments to produce a return value. This trait is designed to be used with OCaml values and provides methods for calling the function, describing its arguments, and generating unique IDs for the function signature.

Type Parameters

Ret: The return type of the function, which must implement ocaml::FromValue and OCamlDesc.

Required Methods

fn call_with(&self, gc: &Runtime, func: Value) -> Ret

fn describe_args(env: &Env, generics: &[&str]) -> Vec<String>

Describes the arguments (i.e. calls OCamlDesc::ocaml_desc) of the function. This method should be provided by downstream trait implementations.

fn unique_id_args() -> Vec<u128>

Generates unique IDs for the function arguments. This method should be provided by downstream trait implementations.

Provided Methods

fn ocaml_desc(env: &Env, generics: &[&str]) -> String

ocaml_desc generates OCaml type signature for this Callable

fn unique_id() -> u128

fn process_result(&self, res: Result<Value, Error>) -> Ret

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.