ocaml_rs_smartptr::func

Struct OCamlFunc

pub struct OCamlFunc<Args, Ret>(/* private fields */);

OCamlFunc is a wrapper around MlBox that represents an OCaml function.
It holds a reference to the OCaml function and ensures that it is safe to call
from Rust. The PhantomData is used to keep track of the argument and return types.

Implementations

impl<Args, Ret> OCamlFunc<Args, Ret>

pub fn new(gc: &Runtime, v: Value) -> Self

Creates a new OCamlFunc from an OCaml value.
This function takes an OCaml runtime handle to ensure that the operation
is called while the OCaml domain lock is acquired.

impl<Args: Callable<Ret>, Ret> OCamlFunc<Args, Ret>
where Ret: OCamlDesc + FromValue,

pub fn call(&self, gc: &Runtime, args: Args) -> Ret

Calls the OCaml function with the provided arguments.
This function ensures that the OCaml runtime is properly handled.