ocaml_rs_smartptr

Module ml_box

This module provides a safe wrapper around ocaml::Value to allow sending OCaml values between threads in Rust. The MlBox struct ensures that the OCaml runtime handle is available when recovering the value, maintaining safety.

Structs

MlBox

This wrapper type around ocaml::root::Root allows sending ocaml::root::Root to other threads and provides a safe interface for doing so - the value can be recovered back only in a context where an OCaml runtime handle is available. As cloning of ocaml::root::Root is not safe outside of the OCaml Domain lock (see comments in the as_value method below), we wrap it with Arc to enable safe cloning of MlBox from Rust.