ocaml_rs_smartptr

Module ocaml_gen_extras

This module provides additional utilities and extensions for generating OCaml bindings.

Structs

OcamlGenPlugin

Represents a plugin for generating OCaml bindings. It contains a generator function and the name of the crate.

P1

P1 is for a single type parameter ’x where x is const C: char

P2

P2 is for a two type parameters ’x,’y where x is const C1: char and y is const C2: char

P3

Same as P2 but for three type parameters

PolymorphicValue

A wrapper around ocaml::Value that is printed by ocaml_gen as an OCaml polymorphic type, i.e., 'a or 'b, where the a or b symbol comes from the const C: char of this PolymorphicValue.

WithTypeParams

Thin wrapper around T which adds ability to print T into ocaml_desc as a type with type parameters

Traits

TypeParams

A trait that is implemented by P1, P2, etc., used as a link between concrete P1, P2, etc., and the WithTypeParams wrapper type below.

Functions

stubs_gen_main

Main function for stubs generation binaries. It collects OcamlGenPlugins registered in other libraries and writes one .ml file per crate with generated OCaml bindings.