ocaml_rs_smartptr

Module ptr

This module provides the DynBox smart pointer, which is a wrapper around the registry’s DynArc with PhantomData for type safety. DynBox allows the user to wrap the object in a Mutex or shared RwLock. By default, using .into() will create a Mutex-protected version (exclusive).

Key Components

DynBox<T>: A smart pointer for dynamically typed Rust objects referenced from the OCaml side.

RustyObj: A thin wrapper around a pointer to DynArc.

Usage

Creating a DynBox

You can create a DynBox with either a Mutex or RwLock:

use ocaml_rs_smartptr::ptr::DynBox;
let exclusive_box = DynBox::new_exclusive(42); // Mutex-protected
let shared_box = DynBox::new_shared("foo"); // RwLock-protected

Coercion

The coerce and coerce_mut methods return a handle that holds a lock. Be cautious to avoid deadlocks when using these methods.

OCaml Integration

DynBox integrates with the ocaml_gen package by providing OCamlDesc and OCamlBinding traits. Polymorphic variants are used to encode supported traits of the Rust type that we wrap/bind.

Example:

module Animal = struct 
  type nonrec t = [ `Ocaml_rs_smartptr_test_stubs_animal_proxy|`Core_marker_send ] Ocaml_rs_smartptr.Rusty_obj.t
  external name : t -> string = "animal_name"
  external noise : t -> string = "animal_noise"
  external talk : t -> unit = "animal_talk"
end

module Sheep = struct 
  type nonrec t = [ `Ocaml_rs_smartptr_test_stubs_sheep|`Core_marker_sync|`Core_marker_send|`Ocaml_rs_smartptr_test_stubs_animal_proxy ] Ocaml_rs_smartptr.Rusty_obj.t
  external create : string -> t = "sheep_create"
  external is_naked : t -> bool = "sheep_is_naked"
  external sheer : t -> unit = "sheep_sheer"
end

module Wolf = struct 
  type nonrec t = [ `Ocaml_rs_smartptr_test_stubs_wolf|`Core_marker_sync|`Core_marker_send|`Ocaml_rs_smartptr_test_stubs_animal_proxy ] Ocaml_rs_smartptr.Rusty_obj.t
  external create : string -> t = "wolf_create"
  external set_hungry : t -> bool -> unit = "wolf_set_hungry"
end

This allows passing Wolf or Sheep whenever Animal is required by using coercion operator in OCaml (:>).

RustyObj

RustyObj is a thin wrapper around a pointer to DynArc. We convert Arc into a raw pointer to hold that raw pointer in the OCaml heap, ensuring that moving of that value by the OCaml GC does not affect any Rust invariants. Reverse operation reconstructs the Arc from the raw pointer. This ensures that both OCaml and Rust always hold valid Arc-baked references to objects they need.

Structs

DynBox

A smart pointer around the registry’s DynArc with PhantomData for type safety. Allows the user to wrap the object in a Mutex or shared RwLock. By default, using .into() will create a Mutex-protected version (exclusive).