ocaml_rs_smartptr::ptr

Struct DynBox

pub struct DynBox<T>
where
    T: Send + ?Sized,
{ /* private fields */ }

A smart pointer around the registryโs DynArc with PhantomData for type safety. Allows the user to wrap the object in a Mutex or shared RwLock. By default, using .into() will create a Mutex-protected version (exclusive).

Implementations

impl<T: 'static + Send> DynBox<T>

pub fn new_exclusive(value: T) -> Self

Creates a DynBox with a Mutex.

Parameters

value: The value to be wrapped in the DynBox.

Returns

A new DynBox instance with Mutex protection.

impl<T: 'static + Send + ?Sized> DynBox<T>

pub fn new_exclusive_boxed(value: Box<T>) -> Self

Creates a DynBox with a Mutex out of a Boxโed T. Useful if T is unsized, e.g. a dyn Trait.

Parameters

value: The value (inside a Box) to be wrapped in the DynBox.

Returns

A new DynBox instance with Mutex protection.

impl<T: 'static + Sync + Send> DynBox<T>

pub fn new_shared(value: T) -> Self

Creates a DynBox with a RwLock.

Parameters

value: The value to be wrapped in the DynBox.

Returns

A new DynBox instance with RwLock protection.

impl<T: 'static + Sync + Send + ?Sized> DynBox<T>

pub fn new_shared_boxed(value: Box<T>) -> Self

Creates a DynBox with a RwLock out of a Boxโed T. Useful if T is unsized, e.g. a dyn Trait.

Parameters

value: The value (inside a Box) to be wrapped in the DynBox.

Returns

A new DynBox instance with RwLock protection.

impl<T: 'static + Send + ?Sized> DynBox<T>

pub fn coerce(&self) -> Handle<T>

Coerces the DynBox to a handle of the specified type.

Returns

A handle to the coerced type. Note that this handle holds a lock, so use with care to avoid deadlocks.

pub fn coerce_mut(&self) -> HandleMut<T>

Coerces the DynBox to a mutable handle of the specified type.

Returns

A mutable handle to the coerced type. Note that this handle holds a lock, so use with care to avoid deadlocks.