ocaml_rs_smartptr

Module registry

This module provides a Registry for managing type coercions and type information in a Rust-OCaml binding context. The Registry allows for the registration of coercion functions that convert between different types, leveraging Rust’s type system and OCaml’s garbage collector.

The primary goal is to facilitate the interaction between Rust and OCaml by providing a mechanism to register and retrieve type information and coercion functions. This is particularly useful when dealing with dynamic types (dyn Any) and ensuring that the OCaml garbage collector can properly manage Rust objects.

The Registry uses TypeId to uniquely identify types and DynArc to store dynamically typed values. Coercion functions are stored in a HashMap and can be retrieved to convert between registered types.

See relevant discussion: https://users.rust-lang.org/t/rust-ocaml-bindings-and-traits/113263 Special thanks to Kevin Reid (https://users.rust-lang.org/u/kpreid) for providing the basis for building this module.

Structs

Plugin

The Plugin struct represents a plugin with an initializer function.

TypeInfo

A struct representing type information, including the fully qualified name and a list of implementations.

Functions

coerce

Coerces a DynArc input to a handle of the specified output type using the global registry.

coerce_mut

Coerces a DynArc input to a mutable handle of the specified output type using the global registry.

get_type_info

Retrieves the type information for a given input type from the global registry.

initialize_plugins

Initializes all registered plugins. This function is called once.

register

Registers coercion functions for converting between types In and Out in the global registry.

register_type

Registers a type in the global registry.

register_type_info

Registers type information in the global registry.

Type Aliases

Handle

A type alias for a handle to a read-only reference of type Out. This is used to represent coerced values in the registry.

HandleMut

A type alias for a handle to a mutable reference of type Out. This is used to represent coerced mutable values in the registry.